0.1.3+rev62

+++ include/uart.h
@@ -20,7 +20,7 @@ static inline void uart_puts(char *s)
     }
 }
 
-static inline void uart_puthex(u32 val)
+static inline void uart_puthex(u64 val)
 {
     char buf[17];
     for (int i = 0; i < 17; i++)
